Remove trailing slash for OpenAPI spec
The OpenAPI generated client in Swift doesn't workr when the base
URL contains a trailing slash since it concatenates the base URL
and the path to the resource without checking for a trailing
slash. The URL contains *two* slashes which makes the load
balancers unhappy. (like https://api.lab5e.com//span/collections)